requirements.txt
click==7.1.2
docx-mailmerge==0.5.0
lxml==4.6.2
PyYAML==5.4.1
